Refactor code for modularity, update gitignore
