+++ .gitignore
@@ -1,3 +1,4 @@
 /build
 /app/build
 /.gradle
+.idea/
