+++ ChangeLog
@@ -1,5 +1,7 @@
 Version	- changes
 
+1.0.10   - More coding style changes @midopa
+-----
 @Nanolx
 1.0.9	- coding style changes
 	- support for Amazon Flex
